Feat: 계층별 권한 부여 완료

+++ server/src/main/java/com/lucky7/preproject/answer/service/AnswerService.java
@@ -4,7 +4,9 @@ import com.lucky7.preproject.answer.entity.Answer;
 import com.lucky7.preproject.answer.repository.AnswerRepository;
 import com.lucky7.preproject.question.entity.Question;
 import com.lucky7.preproject.question.repository.QuestionRepository;
+import com.lucky7.preproject.user.entity.User;
 import lombok.AllArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,10 +33,13 @@ public class AnswerService {
         }
     }
 
-    public  Answer updateAnswer(long questionId, long answerId, Answer answerToUpdate) {
+    public  Answer updateAnswer(long questionId, long answerId, Answer answerToUpdate, User user) {
         // 수정 요청받은 questionId와 answerId로 해당 Answer를 찾습니다.
         Answer existingAnswer = answerRepository.findById(answerId).orElse(null);
 
+        if (!existingAnswer.getUser().equals(user)) {
+            throw new AccessDeniedException("You do not have permission to update this question.");
+        }
         // Answer 가 존재하지 않을 경우, 즉 존재하지 않는 answerId를 요청했을 경우
         if (existingAnswer == null || existingAnswer.getQuestion().getQuestionId() != questionId) {
             return null;
@@ -61,11 +66,14 @@ public class AnswerService {
         return answerRepository.findById(answerId).orElse(defaultAnswer);
     }
 
-    public Answer deleteAnswer(long questionId, long answerId) {
+    public Answer deleteAnswer(long questionId, long answerId, User user) {
         Answer existingAnswer = answerRepository.findById(answerId).orElse(null);
 
         if (existingAnswer == null || existingAnswer.getQuestion().getQuestionId() != questionId) {
             return null;
+        }
+        if (!existingAnswer.getUser().equals(user)) {
+            throw new AccessDeniedException("You do not have permission to update this answer.");
         }
 
         answerRepository.delete(existingAnswer);
